requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
eventlet==0.33.0  # Required for Flask-SocketIO if using async_mode='eventlet'
dnspython==2.2.1
flask-cors
